add links to privacy policy and legal disclosure
This commit adds links to the privacy policy and legal disclosure pages,
as required by the European General Data Protection Regulation (GDPR)
and Section 5 of the Telemedia Act (TMG).

The privacy policy provides information on how user data is collected,
processed, and protected. The legal disclosure contains information on
the website owner's identity and contact details, as well as any legal
requirements for operating the website.

Link: https://eur-lex.europa.eu/legal-content/EN/TXT/PDF/?uri=CELEX:32016R0679&from=EN
Link: https://www.gesetze-im-internet.de/tmg/__5.html
Signed-off-by: Stefan Kühnel <git@stefankuehnel.com>