workflows: Migrate from 'Netlify' to 'webgo'
Currently it is not possible to deploy a page using the Netlify CLI
via GitHub Actions. This is due to the following error:

  Multiple possible build commands found
   ›   Error: Detected commands for: Hugo, Zola. Update your settings
              to specify which to use. Refer to https://ntl.fyi/dev-monorepo
	      for more information.

A bug report for this was submitted to Netlify's GitHub issue tracker
on 27 September 2024 [1]. However, a fix has not yet been provided.

Therefore, migrate from Netlify back to webgo.

Link: https://github.com/netlify/cli/issues/6841 [1]
Signed-off-by: Stefan Kühnel <git@stefankuehnel.com>